# Optimize qdigest histogram, allow for custom long-to-float conversion for middle