@tdcmeehan 2018-11-05
Thanks for the comments @martint.  I addressed your feedback, can you please give it another pass?